# ls10/api-specification.yaml
openapi: 3.0.0
info:
  title: TeachingX API
  version: 1.0.0
  description: Complete API documentation for TeachingX - Users, Posts, Authentication, and File Upload
  contact:
    name: API Support
    email: support@teachingx.com
  license:
    name: ISC

servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:3000
    description: Local server

tags:
  - name: Authentication
    description: User authentication endpoints
  - name: Users
    description: User management endpoints
  - name: Posts
    description: Post management endpoints
  - name: Upload
    description: File upload endpoints

paths:
  /health:
    get:
      summary: Health check endpoint
      tags:
        - System
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                example:
                  success: true
                  message: "API is running"

  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  example: john_doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: password123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already exists

  /api/auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                    description: JWT token
                  user:
                    $ref: '#/components/schemas/User'
                example:
                  success: true
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  user:
                    _id: "507f1f77bcf86cd799439011"
                    username: "john_doe"
                    email: "john@example.com"
        '401':
          description: Invalid email or password
        '400':
          description: Bad request

  /api/users:
    get:
      summary: Get all users
      tags:
        - Users
      description: Retrieve a list of all users in the system
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
          description: Number of results to return
        - in: query
          name: skip
          schema:
            type: integer
            example: 0
          description: Number of results to skip
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  total:
                    type: integer
                example:
                  success: true
                  data:
                    - _id: "507f1f77bcf86cd799439011"
                      username: "john_doe"
                      email: "john@example.com"
                      avatar: "https://example.com/avatar.jpg"
                      createdAt: "2024-01-15T10:30:00Z"
                  total: 1
        '500':
          description: Server error

    post:
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
              properties:
                username:
                  type: string
                  example: jane_doe
                email:
                  type: string
                  format: email
                  example: jane@example.com
                avatar:
                  type: string
                  example: "https://example.com/avatar.jpg"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      summary: Get user by ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
          description: User MongoDB ObjectId
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid user ID format
        '404':
          description: User not found

    put:
      summary: Update user by ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: jane_doe_updated
                email:
                  type: string
                  format: email
                  example: jane_updated@example.com
                avatar:
                  type: string
                  example: "https://example.com/new-avatar.jpg"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Bad request or invalid ID
        '404':
          description: User not found

    delete:
      summary: Delete user by ID
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid user ID format
        '404':
          description: User not found

  /api/posts:
    get:
      summary: Get all posts
      tags:
        - Posts
      description: Retrieve all posts with pagination support
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            example: 20
        - in: query
          name: skip
          schema:
            type: integer
            example: 0
        - in: query
          name: author
          schema:
            type: string
            format: ObjectId
          description: Filter by author ID
      responses:
        '200':
          description: List of all posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  total:
                    type: integer
        '500':
          description: Server error

    post:
      summary: Create a new post
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
                - author
              properties:
                title:
                  type: string
                  example: "My First Post"
                content:
                  type: string
                  example: "This is the content of my first post"
                author:
                  type: string
                  format: ObjectId
                  example: "507f1f77bcf86cd799439011"
                image:
                  type: string
                  example: "https://example.com/post-image.jpg"
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/posts/{id}:
    get:
      summary: Get post by ID
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid post ID format
        '404':
          description: Post not found

    put:
      summary: Update post by ID
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Post Title"
                content:
                  type: string
                  example: "Updated post content"
                image:
                  type: string
                  example: "https://example.com/updated-image.jpg"
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Bad request or invalid ID
        '404':
          description: Post not found

    delete:
      summary: Delete post by ID
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: ObjectId
      responses:
        '200':
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Invalid post ID format
        '404':
          description: Post not found

  /api/upload:
    post:
      summary: Upload a single image
      tags:
        - Upload
      description: Upload a single image file (max 5MB, supports jpg, png, gif, webp)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      url:
                        type: string
                        description: Cloudinary URL
                      public_id:
                        type: string
                        description: Cloudinary public ID
                example:
                  success: true
                  message: "Image uploaded successfully"
                  data:
                    url: "https://res.cloudinary.com/..."
                    public_id: "teaching-x/..."
        '400':
          description: No file uploaded or invalid file
        '413':
          description: File too large
        '415':
          description: Unsupported media type

  /api/upload/multiple:
    post:
      summary: Upload multiple images
      tags:
        - Upload
      description: Upload multiple image files at once (max 5 files, 5MB each)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Multiple image files to upload
      responses:
        '200':
          description: Images uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        url:
                          type: string
                        public_id:
                          type: string
                example:
                  success: true
                  message: "Images uploaded successfully"
                  data:
                    - url: "https://res.cloudinary.com/..."
                      public_id: "teaching-x/..."
                    - url: "https://res.cloudinary.com/..."
                      public_id: "teaching-x/..."
        '400':
          description: No files uploaded or validation error
        '413':
          description: File(s) too large
        '415':
          description: Unsupported media type

components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
          description: User unique identifier
        username:
          type: string
          description: User's username
        email:
          type: string
          format: email
          description: User's email address
        avatar:
          type: string
          description: URL to user's avatar image
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        _id: "507f1f77bcf86cd799439011"
        username: "john_doe"
        email: "john@example.com"
        avatar: "https://example.com/avatar.jpg"
        createdAt: "2024-01-15T10:30:00Z"
        updatedAt: "2024-01-15T10:30:00Z"

    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/User'
      example:
        success: true
        message: "User created successfully"
        data:
          _id: "507f1f77bcf86cd799439011"
          username: "john_doe"
          email: "john@example.com"
          avatar: "https://example.com/avatar.jpg"
          createdAt: "2024-01-15T10:30:00Z"
          updatedAt: "2024-01-15T10:30:00Z"

    Post:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
          description: Post unique identifier
        title:
          type: string
          description: Post title
        content:
          type: string
          description: Post content/body
        author:
          type: string
          format: ObjectId
          description: Author's user ID
        image:
          type: string
          description: URL to post's main image
        likes:
          type: integer
          description: Number of likes
        comments:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
          description: Post creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      example:
        _id: "507f1f77bcf86cd799439012"
        title: "My First Post"
        content: "This is the content of my first post"
        author: "507f1f77bcf86cd799439011"
        image: "https://example.com/post-image.jpg"
        likes: 5
        comments: []
        createdAt: "2024-01-15T10:30:00Z"
        updatedAt: "2024-01-15T10:30:00Z"

    PostResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Post'
      example:
        success: true
        message: "Post created successfully"
        data:
          _id: "507f1f77bcf86cd799439012"
          title: "My First Post"
          content: "This is the content of my first post"
          author: "507f1f77bcf86cd799439011"
          image: "https://example.com/post-image.jpg"
          likes: 0
          comments: []
          createdAt: "2024-01-15T10:30:00Z"
          updatedAt: "2024-01-15T10:30:00Z"

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation failed"
        errors:
          type: object
          description: Detailed validation errors
      example:
        success: false
        message: "Validation failed"
        errors:
          email: "Email is required"
          username: "Username must be at least 3 characters"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication

security:
  - bearerAuth: []